googleapis/requirements.txt
google-api-python-client
PyDrive==1.3.1
sshtunnel==0.4.0
mysql-connector-python==9.5.0
paramiko==3.4.0
tqdm
tenacity
